move matching files serialization tests to kernel tests
