+++ tensorflow/python/data/experimental/kernel_tests/serialization/BUILD
@@ -242,22 +242,6 @@ tf_py_test(
 )
 
 tf_py_test(
-    name = "matching_files_dataset_serialization_test",
-    size = "small",
-    srcs = ["matching_files_dataset_serialization_test.py"],
-    tags = [
-        "no_windows",
-    ],
-    deps = [
-        "//tensorflow/python:client_testlib",
-        "//tensorflow/python/data/experimental/ops:matching_files",
-        "//tensorflow/python/data/kernel_tests:checkpoint_test_base",
-        "//tensorflow/python/data/ops:dataset_ops",
-        "//third_party/py/numpy",
-    ],
-)
-
-tf_py_test(
     name = "optimize_dataset_serialization_test",
     size = "medium",
     srcs = ["optimize_dataset_serialization_test.py"],
